# [Refactor] Add ExploreGroupTask in memo phase