--- a/fe/fe-core/src/test/java/com/starrocks/sql/plan/ScanTest.java
+++ b/fe/fe-core/src/test/java/com/starrocks/sql/plan/ScanTest.java
@@ -139,7 +139,7 @@ public class ScanTest extends PlanTestBase {
         String sql = "select v1,v2,v3 from t0";
         connectContext.getSessionVariable().setSingleNodeExecPlan(true);
         String plan = getFragmentPlan(sql);
-        Assert.assertTrue(plan.contains("PLAN FRAGMENT 0\n" +
+        assertContains(plan, "PLAN FRAGMENT 0\n" +
                 " OUTPUT EXPRS:1: v1 | 2: v2 | 3: v3\n" +
                 "  PARTITION: RANDOM\n" +
                 "\n" +
@@ -154,7 +154,7 @@ public class ScanTest extends PlanTestBase {
                 "     tabletList=\n" +
                 "     cardinality=1\n" +
                 "     avgRowSize=3.0\n" +
-                "     numNodes=0"));
+                "     numNodes=0");
         connectContext.getSessionVariable().setSingleNodeExecPlan(false);
     }
 
@@ -209,12 +209,12 @@ public class ScanTest extends PlanTestBase {
         sql = "select t1.k2, sum(t1.k9) from baseall t1 " +
                 "join join2 t2 on t1.k1 = t2.id join baseall t3 on t1.k1 = t3.k1 group by t1.k2";
         plan = getFragmentPlan(sql);
-        assertContains(plan, "  |----2:OlapScanNode\n" +
-                "  |       TABLE: baseall\n" +
-                "  |       PREAGGREGATION: OFF. Reason: Has can not pre-aggregation Join");
-        assertContains(plan, "0:OlapScanNode\n" +
+        assertContains(plan, "  0:OlapScanNode\n" +
                 "     TABLE: baseall\n" +
-                "     PREAGGREGATION: ON");
+                "     PREAGGREGATION: ON\n");
+        assertContains(plan, "  2:OlapScanNode\n" +
+                "     TABLE: join2\n" +
+                "     PREAGGREGATION: ON\n");
 
         sql = "select t3.k2, sum(t3.k9) from baseall t1 " +
                 "join [broadcast] join2 t2 on t1.k1 = t2.id join [broadcast] baseall t3 " +