… fe/fe-core/src/test/resources/sql/external/hive/tpch/q18.sql
… fe/fe-core/src/test/resources/sql/external/hive/tpch/q9.sql
… fe/fe-core/src/test/resources/sql/tpch-histogram-cost/q9.sql
… fe/fe-core/src/test/resources/sql/tpch/q21.sql
… fe/fe-core/src/test/resources/sql/tpchcost/q9.sql