@wanpengfei-git 2022-08-25
#### [FE PR Coverage Check]

:disappointed: **fail** : 17 / 43 (39.53%)




#### file detail

|   |path|covered_line|new_line|coverage|not_covered_line_detail|
|----|----|----|----|----|----|
|:large_blue_circle:|com/starrocks/sql/optimizer/task/ExploreGroupTask.java|0|11|00.00%|[15, 16, 17, 21, 22, 25, 26, 27, 29, 30, 34]|
|:large_blue_circle:|com/starrocks/sql/optimizer/task/DeriveStatsTask.java|2|8|25.00%|[33, 34, 35, 46, 47, 48]|
|:large_blue_circle:|com/starrocks/sql/optimizer/Group.java|2|6|33.33%|[77, 81, 82, 191]|
|:large_blue_circle:|com/starrocks/sql/optimizer/task/ApplyRuleTask.java|1|2|50.00%|[92]|
|:large_blue_circle:|com/starrocks/sql/optimizer/task/OptimizeExpressionTask.java|11|15|73.33%|[36, 37, 38, 70]|
|:large_blue_circle:|com/starrocks/sql/optimizer/task/SeriallyTaskScheduler.java|1|1|100.00%|[]|