@kangkaisen 2022-08-29
Need full performance test and optimizer itself performance test.