@Seaven 2022-08-30
> Need full performance test and optimizer itself performance test.

http://39.101.206.100:8090/view/benchmark/job/performance_comparison/326/Two_20version_20compare_20report/