@Seaven 2022-08-30
run starrocks_admit_test