@wanpengfei-git 2022-09-02
run starrocks_admit_test