@Seaven 2022-09-03
run starrocks_admit_test