@Seaven 2022-09-03
run starrocks_fe_unittest