@wanpengfei-git 2022-09-03
#### [FE PR Coverage Check]

:heart_eyes: **pass** : 44 / 47 (93.62%)




#### file detail

|   |path|covered_line|new_line|coverage|not_covered_line_detail|
|----|----|----|----|----|----|
|:large_blue_circle:|com/starrocks/sql/optimizer/task/ApplyRuleTask.java|2|3|66.67%|[49]|
|:large_blue_circle:|com/starrocks/sql/optimizer/task/ExploreGroupTask.java|10|11|90.91%|[34]|
|:large_blue_circle:|com/starrocks/sql/optimizer/task/OptimizeExpressionTask.java|15|16|93.75%|[57]|
|:large_blue_circle:|com/starrocks/sql/optimizer/task/DeriveStatsTask.java|7|7|100.00%|[]|
|:large_blue_circle:|com/starrocks/sql/optimizer/Group.java|6|6|100.00%|[]|
|:large_blue_circle:|com/starrocks/sql/optimizer/task/OptimizeGroupTask.java|4|4|100.00%|[]|